platform: tegra210: rework ptsa
Existing PTSA calculations expect hardcoded DDA values in emc tables
and in this driver, with a maximum of 2132 MHz emc rate.

This is not enough when that rate exceeds it and ptsa rate becomes higher than
1.499. Also the current values reach the 1.999 absolute max when emc rate
exceeds 2841 MHz.

So rework the driver to dynamically change max PTSA DDA when emc rate exceeds
hardcoded max and ptsa rework is enabled.

The result of the reworked PTSA changes is that the memory controller's
snap arbiter can properly prioritize work on high emc rates, increasing
efficiency significantly on bandwidth pressured scenarios.

A new module parameter was added to enable that: `latency_allowance.ptsa_rework`